Minimum player config

+++ mafia.js
@@ -386,31 +386,35 @@ var baseCommands = [
             if (gameInChannel) {
                 if (gameInChannel.hostId == message.author.id) {
                     if (gameInChannel.state == STATE.INIT) {
-                        mafiabot.createChannel(message.channel, 'mafia' + Math.random().toString().substring(2), 'text', (error, mafiaChannel) => {
-                            if (mafiaChannel) {
-                                gameInChannel.state = STATE.CONFIRMING;
-                                gameInChannel.mafiaChannelId = mafiaChannel.id;
-                                mafiabot.syncMessage(message.channel.id, `Sending out roles for game of mafia hosted by <@${gameInChannel.hostId}>! Check your PMs for info and type **${pre}confirm** in this channel to confirm your role.`);
-                                printCurrentPlayers(message.channel.id);
-                                for (var i = 0; i < gameInChannel.players.length; i++) {
-                                    var player = gameInChannel.players[i];
-                                    player.faction = ['Town', 'Mafia'][Math.floor(Math.random() * 2)];
-                                    player.role = roles[Math.floor(Math.random() * roles.length)].id;
-                                    mafiabot.sendMessage(_.find(mafiabot.users, {id: player.id}), `Your role is ***${player.faction} ${getRole(player.role).name}***.\n${getRole(player.role).description}\nType **${pre}confirm** in <#${message.channel.id}> to confirm your participation in the game of mafia hosted by <@${gameInChannel.hostId}>.`);
-                                }
+                        if (gameInChannel.players.length >= config.minimumPlayers) {
+                            mafiabot.createChannel(message.channel, 'mafia' + Math.random().toString().substring(2), 'text', (error, mafiaChannel) => {
+                                if (mafiaChannel) {
+                                    gameInChannel.state = STATE.CONFIRMING;
+                                    gameInChannel.mafiaChannelId = mafiaChannel.id;
+                                    mafiabot.syncMessage(message.channel.id, `Sending out roles for game of mafia hosted by <@${gameInChannel.hostId}>! Check your PMs for info and type **${pre}confirm** in this channel to confirm your role.`);
+                                    printCurrentPlayers(message.channel.id);
+                                    for (var i = 0; i < gameInChannel.players.length; i++) {
+                                        var player = gameInChannel.players[i];
+                                        player.faction = ['Town', 'Mafia'][Math.floor(Math.random() * 2)];
+                                        player.role = roles[Math.floor(Math.random() * roles.length)].id;
+                                        mafiabot.sendMessage(_.find(mafiabot.users, {id: player.id}), `Your role is ***${player.faction} ${getRole(player.role).name}***.\n${getRole(player.role).description}\nType **${pre}confirm** in <#${message.channel.id}> to confirm your participation in the game of mafia hosted by <@${gameInChannel.hostId}>.`);
+                                    }
 
-                                var everyoneId = _.find(mafiaChannel.server.roles, {name: "@everyone"}).id;
-                                var mafiaPlayers = gameInChannel.players;//_.filter(gameInChannel.players, {faction: 'Mafia'});
-                                mafiabot.overwritePermissions(mafiaChannel, everyoneId, { readMessages: false, sendMessages: false });
-                                for (var i = 0; i < mafiaPlayers.length; i++) {
-                                    var mafiaPlayer = _.find(mafiabot.users, {id: mafiaPlayers[i].id});
-                                    mafiabot.overwritePermissions(mafiaChannel, mafiaPlayer, { readMessages: true, sendMessages: true });
-                                    mafiabot.sendMessage(mafiaPlayer, `Use the channel <#${mafiaChannel.id}> to chat with your fellow Mafia team members, and to send in your nightly kill.`);
+                                    var everyoneId = _.find(mafiaChannel.server.roles, {name: "@everyone"}).id;
+                                    var mafiaPlayers = gameInChannel.players;//_.filter(gameInChannel.players, {faction: 'Mafia'});
+                                    mafiabot.overwritePermissions(mafiaChannel, everyoneId, { readMessages: false, sendMessages: false });
+                                    for (var i = 0; i < mafiaPlayers.length; i++) {
+                                        var mafiaPlayer = _.find(mafiabot.users, {id: mafiaPlayers[i].id});
+                                        mafiabot.overwritePermissions(mafiaChannel, mafiaPlayer, { readMessages: true, sendMessages: true });
+                                        mafiabot.sendMessage(mafiaPlayer, `Use the channel <#${mafiaChannel.id}> to chat with your fellow Mafia team members, and to send in your nightly kill.`);
+                                    }
+                                    mafiabot.syncMessage(mafiaChannel.id, `**Welcome to the mafia team!**\nYour team is:${listUsers(_.map(mafiaPlayers, 'id'))}`);
+                                    mafiabot.syncMessage(mafiaChannel.id, `As a team you have **1 kill each night**. Use the ***${pre}kill*** command to use that ability when I prompt you in this chat.`);
                                 }
-                                mafiabot.syncMessage(mafiaChannel.id, `**Welcome to the mafia team!**\nYour team is:${listUsers(_.map(mafiaPlayers, 'id'))}`);
-                                mafiabot.syncMessage(mafiaChannel.id, `As a team you have **1 kill each night**. Use the ***${pre}kill*** command to use that ability when I prompt you in this chat.`);
-                            }
-                        });
+                            });
+                        } else {
+                            mafiabot.reply(message, `You need a minimum of ${s(config.minimumPlayers, 'player')} to start a game!`);
+                        }
                     } else if (gameInChannel.state == STATE.READY) {
                         gameInChannel.state = STATE.DAY;
                         gameInChannel.day = 1;
